requirements.txt
PyYAML==5.2
aiofiles==0.4.0
coverage==5.0.1
